refactor: split fastify instance in different files to run tests
