+++ package.json
@@ -1,125 +1,125 @@
 {
-  "name": "ts-cat",
-  "type": "module",
-  "version": "1.2.5",
-  "private": true,
-  "packageManager": "pnpm@8.10.2",
-  "description": "TypeScript version of the Cheshire Cat AI project",
-  "author": "zAlweNy26 & Contributors",
-  "license": "GPL-3.0-only",
-  "homepage": "https://github.com/zAlweNy26/ts-cat",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/zAlweNy26/ts-cat.git"
-  },
-  "bugs": {
-    "url": "https://github.com/zAlweNy26/ts-cat/issues"
-  },
-  "keywords": [
-    "fastify",
-    "typescript",
-    "ai",
-    "javascript",
-    "cheshire-cat",
-    "langchain",
-    "assistant",
-    "framework"
-  ],
-  "main": "dist/main.js",
-  "module": "dist/main.mjs",
-  "types": "dist/main.d.ts",
-  "files": [
-    ".d.ts",
-    "dist"
-  ],
-  "engines": {
-    "node": ">=18.0.0"
-  },
-  "scripts": {
-    "build": "tsup src/** --format cjs,esm --dts",
-    "lint": "eslint . --fix",
-    "test": "vitest dev --run",
-    "test:ui": "vitest dev --ui",
-    "release": "pnpm run lint && pnpm run test && changelogen --release && git push --follow-tags",
-    "dev": "pnpm run \"/dev:/\"",
-    "dev:tsc": "tsc --watch --preserveWatchOutput",
-    "dev:node": "WATCH=true pnpm start",
-    "start": "tsx watch --no-warnings=ExperimentalWarning ./index.ts"
-  },
-  "dependencies": {
-    "@fastify/cors": "^9.0.1",
-    "@fastify/formbody": "^7.4.0",
-    "@fastify/multipart": "^8.2.0",
-    "@fastify/sensible": "^5.5.0",
-    "@fastify/static": "^7.0.3",
-    "@fastify/swagger": "^8.14.0",
-    "@fastify/swagger-ui": "^3.0.0",
-    "@fastify/under-pressure": "^8.3.0",
-    "@fastify/websocket": "^9.0.0",
-    "@langchain/anthropic": "^0.1.11",
-    "@langchain/azure-openai": "^0.0.6",
-    "@langchain/cohere": "^0.0.6",
-    "@langchain/community": "^0.0.45",
-    "@langchain/core": "^0.1.55",
-    "@langchain/google-genai": "^0.0.10",
-    "@langchain/mistralai": "^0.0.16",
-    "@langchain/openai": "^0.0.26",
-    "@mgcrea/fastify-request-logger": "^1.5.0",
-    "@mgcrea/pino-pretty-compact": "^1.3.0",
-    "@qdrant/js-client-rest": "^1.8.2",
-    "callsites": "^4.1.0",
-    "cheerio": "1.0.0-rc.12",
-    "chokidar": "^3.6.0",
-    "consola": "^3.2.3",
-    "console-table-printer": "^2.12.0",
-    "d3-dsv": "^2.0.0",
-    "date-fns": "^3.6.0",
-    "defu": "^6.1.4",
-    "destr": "^2.0.3",
-    "dotenv": "^16.4.5",
-    "fastembed": "^1.14.1",
-    "fastify": "^4.26.2",
-    "fastify-zod-openapi": "^1.2.0",
-    "get-port-please": "^3.1.2",
-    "html-to-text": "^9.0.5",
-    "is-docker": "^3.0.0",
-    "js-tiktoken": "^1.0.10",
-    "langchain": "^0.1.32",
-    "lodash": "^4.17.21",
-    "lowdb": "^7.0.1",
-    "mammoth": "^1.7.1",
-    "ofetch": "^1.3.4",
-    "officeparser": "^4.0.8",
-    "pdf-parse": "^1.1.1",
-    "pkg-types": "^1.0.3",
-    "qs": "^6.12.0",
-    "scule": "^1.3.0",
-    "turbowatch": "^2.29.4",
-    "uncrypto": "^0.1.3",
-    "zod": "^3.22.4",
-    "zod-openapi": "^2.15.0",
-    "zod-to-json-schema": "^3.22.5"
-  },
-  "devDependencies": {
-    "@antfu/eslint-config": "^2.13.2",
-    "@total-typescript/ts-reset": "^0.5.1",
-    "@types/lodash": "^4.17.0",
-    "@types/node": "^20.12.6",
-    "@types/nodemon": "^1.19.6",
-    "@types/qs": "^6.9.14",
-    "@types/ws": "^8.5.10",
-    "@vitest/ui": "^1.4.0",
-    "changelogen": "^0.5.5",
-    "eslint": "^8.57.0",
-    "eslint-plugin-format": "^0.1.0",
-    "eslint-plugin-zod-openapi": "^0.1.0",
-    "magicast": "^0.3.3",
-    "nodemon": "^3.1.0",
-    "tsup": "^8.0.2",
-    "tsx": "^4.7.2",
-    "typescript": "^5.4.4",
-    "typescript-eslint": "^7.6.0",
-    "vite-tsconfig-paths": "^4.3.2",
-    "vitest": "^1.4.0"
-  }
+	"name": "ts-cat",
+	"type": "module",
+	"version": "1.2.5",
+	"private": true,
+	"packageManager": "pnpm@8.10.2",
+	"description": "TypeScript version of the Cheshire Cat AI project",
+	"author": "zAlweNy26 & Contributors",
+	"license": "GPL-3.0-only",
+	"homepage": "https://github.com/zAlweNy26/ts-cat",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/zAlweNy26/ts-cat.git"
+	},
+	"bugs": {
+		"url": "https://github.com/zAlweNy26/ts-cat/issues"
+	},
+	"keywords": [
+		"fastify",
+		"typescript",
+		"ai",
+		"javascript",
+		"cheshire-cat",
+		"langchain",
+		"assistant",
+		"framework"
+	],
+	"main": "dist/server.js",
+	"module": "dist/server.mjs",
+	"types": "dist/server.d.ts",
+	"files": [
+		".d.ts",
+		"dist"
+	],
+	"engines": {
+		"node": ">=18.0.0"
+	},
+	"scripts": {
+		"build": "tsup src/** --format cjs,esm --dts",
+		"lint": "eslint . --fix",
+		"test": "vitest dev --run",
+		"test:ui": "vitest dev --ui",
+		"release": "pnpm run lint && pnpm run test && changelogen --release && git push --follow-tags",
+		"dev": "pnpm run \"/dev:/\"",
+		"dev:tsc": "tsc --watch --preserveWatchOutput",
+		"dev:node": "WATCH=true pnpm start",
+		"start": "tsx watch --no-warnings=ExperimentalWarning ./index.ts"
+	},
+	"dependencies": {
+		"@fastify/cors": "^9.0.1",
+		"@fastify/formbody": "^7.4.0",
+		"@fastify/multipart": "^8.2.0",
+		"@fastify/sensible": "^5.5.0",
+		"@fastify/static": "^7.0.3",
+		"@fastify/swagger": "^8.14.0",
+		"@fastify/swagger-ui": "^3.0.0",
+		"@fastify/under-pressure": "^8.3.0",
+		"@fastify/websocket": "^9.0.0",
+		"@langchain/anthropic": "^0.1.11",
+		"@langchain/azure-openai": "^0.0.6",
+		"@langchain/cohere": "^0.0.6",
+		"@langchain/community": "^0.0.45",
+		"@langchain/core": "^0.1.55",
+		"@langchain/google-genai": "^0.0.10",
+		"@langchain/mistralai": "^0.0.16",
+		"@langchain/openai": "^0.0.26",
+		"@mgcrea/fastify-request-logger": "^1.5.0",
+		"@mgcrea/pino-pretty-compact": "^1.3.0",
+		"@qdrant/js-client-rest": "^1.8.2",
+		"callsites": "^4.1.0",
+		"cheerio": "1.0.0-rc.12",
+		"chokidar": "^3.6.0",
+		"consola": "^3.2.3",
+		"console-table-printer": "^2.12.0",
+		"d3-dsv": "^2.0.0",
+		"date-fns": "^3.6.0",
+		"defu": "^6.1.4",
+		"destr": "^2.0.3",
+		"dotenv": "^16.4.5",
+		"fastembed": "^1.14.1",
+		"fastify": "^4.26.2",
+		"fastify-zod-openapi": "^1.2.0",
+		"get-port-please": "^3.1.2",
+		"html-to-text": "^9.0.5",
+		"is-docker": "^3.0.0",
+		"js-tiktoken": "^1.0.10",
+		"langchain": "^0.1.32",
+		"lodash": "^4.17.21",
+		"lowdb": "^7.0.1",
+		"mammoth": "^1.7.1",
+		"ofetch": "^1.3.4",
+		"officeparser": "^4.0.8",
+		"pdf-parse": "^1.1.1",
+		"pkg-types": "^1.0.3",
+		"qs": "^6.12.0",
+		"scule": "^1.3.0",
+		"turbowatch": "^2.29.4",
+		"uncrypto": "^0.1.3",
+		"zod": "^3.22.4",
+		"zod-openapi": "^2.15.0",
+		"zod-to-json-schema": "^3.22.5"
+	},
+	"devDependencies": {
+		"@antfu/eslint-config": "^2.13.2",
+		"@total-typescript/ts-reset": "^0.5.1",
+		"@types/lodash": "^4.17.0",
+		"@types/node": "^20.12.6",
+		"@types/nodemon": "^1.19.6",
+		"@types/qs": "^6.9.14",
+		"@types/ws": "^8.5.10",
+		"@vitest/ui": "^1.4.0",
+		"changelogen": "^0.5.5",
+		"eslint": "^8.57.0",
+		"eslint-plugin-format": "^0.1.0",
+		"eslint-plugin-zod-openapi": "^0.1.0",
+		"magicast": "^0.3.3",
+		"nodemon": "^3.1.0",
+		"tsup": "^8.0.2",
+		"tsx": "^4.7.2",
+		"typescript": "^5.4.4",
+		"typescript-eslint": "^7.6.0",
+		"vite-tsconfig-paths": "^4.3.2",
+		"vitest": "^1.4.0"
+	}
 }
