+++ test/env-mock.ts
@@ -1,19 +1,10 @@
-import { madHatter } from '@mh/mad-hatter.ts'
-import { cheshireCat } from '@lg/cheshire-cat.ts'
-import { StrayCat } from '@lg/stray-cat.ts'
-import { rabbitHole } from '@rh'
 import { afterAll, beforeAll, test, vi } from 'vitest'
-import app from '@/main.ts'
-
-export const mockTest = test.extend({
-	/* cheshireCat,
-	madHatter,
-	rabbitHole,
-	strayCat: new StrayCat('Alice'), */
-})
+import * as utils from '@utils'
+import app from '@/app.ts'
 
 export const appTest = test.extend({
 	app,
+	utils,
 })
 
 beforeAll(() => {
