+++ test/looking_glass/cheshire-cat.test.ts
@@ -1,6 +1,7 @@
-import { mockTest } from 'test/env-mock.ts'
+import { mockTest } from 'test/class-mocks.ts'
 import { expect } from 'vitest'
 
 mockTest('test main module', () => {
 	expect(1 + 1).toBe(2)
+	// expect(cat).toBeDefined()
 })
