+++ test/mad_hatter/mad-hatter.test.ts
@@ -1,6 +1,10 @@
-import { mockTest } from 'test/env-mock.ts'
+import { mockTest } from 'test/class-mocks.ts'
 import { expect } from 'vitest'
 
-mockTest('test main module', () => {
+mockTest('test main module', async () => {
 	expect(1 + 1).toBe(2)
+
+	/* expect(mh.installedPlugins.length).toBe(0)
+	await mh.installPlugin('src/plugins/test_plugin')
+	expect(mh.installedPlugins.length).toBe(1) */
 })
